chore: update dependencies and refactor ai-chat module structure
- Updated @follow-app/client-sdk to version 0.3.26 in pnpm-lock.yaml and pnpm-workspace.yaml.
- Upgraded ai package to version 5.0.2 across various package.json files.
- Refactored ai-chat components to use new store structure, moving types and hooks from __internal__ to store.
- Introduced new AIChatContext and hooks for better state management in the ai-chat module.

These changes enhance the maintainability and performance of the ai-chat module while ensuring compatibility with updated dependencies.
