+++ apps/desktop/layer/renderer/package.json
@@ -22,7 +22,6 @@
     "@follow/shared": "workspace:*",
     "@follow/store": "workspace:*",
     "@follow/tracker": "workspace:*",
-    "@folo-services/constants": "0.1.6",
     "@fontsource/sn-pro": "5.2.5",
     "@hcaptcha/react-hcaptcha": "1.12.0",
     "@headlessui/react": "2.2.4",
@@ -51,7 +50,7 @@
     "@use-gesture/react": "10.3.1",
     "@welldone-software/why-did-you-render": "10.0.1",
     "@yornaath/batshit": "0.10.1",
-    "ai": "5.0.0",
+    "ai": "5.0.2",
     "camelcase-keys": "9.1.3",
     "class-variance-authority": "0.7.1",
     "clsx": "2.1.1",
@@ -119,7 +118,7 @@
     "@follow/models": "workspace:*",
     "@follow/types": "workspace:*",
     "@follow/utils": "workspace:*",
-    "@folo-services/ai-tools": "0.2.14",
+    "@folo-services/ai-tools": "0.2.16",
     "@types/node": "24.0.10",
     "@vite-pwa/assets-generator": "1.0.0",
     "fake-indexeddb": "6.0.1",
