+++ apps/desktop/layer/renderer/src/modules/ai-chat/store/chat-core/chat-instance.ts
