+++ apps/desktop/layer/renderer/src/modules/ai-chat/store/event-system/event-emitter.ts
