+++ apps/desktop/layer/renderer/src/modules/ai-chat/store/slices/block.slice.ts
