+++ apps/desktop/layer/renderer/src/modules/ai-chat/store/slices/chat.slice.ts
