+++ apps/desktop/layer/renderer/src/modules/entry-content/components/entry-content/EntryContent.ai.tsx
@@ -24,8 +24,8 @@ import { useInPeekModal } from "~/components/ui/modal/inspire/InPeekModal"
 import { HotkeyScope } from "~/constants"
 import { useRouteParamsSelector } from "~/hooks/biz/useRouteParams"
 import { useFeedSafeUrl } from "~/hooks/common/useFeedSafeUrl"
-import { useBlockActions } from "~/modules/ai-chat/__internal__/hooks"
-import { BlockSliceAction } from "~/modules/ai-chat/__internal__/slices/block.slice"
+import { useBlockActions } from "~/modules/ai-chat/store/hooks"
+import { BlockSliceAction } from "~/modules/ai-chat/store/slices/block.slice"
 import { COMMAND_ID } from "~/modules/command/commands/id"
 
 import { ApplyEntryActions } from "../../ApplyEntryActions"
