+++ apps/desktop/layer/renderer/src/modules/entry-content/components/layouts/ArticleLayout.tsx
@@ -14,8 +14,8 @@ import { useInPeekModal } from "~/components/ui/modal/inspire/InPeekModal"
 import { readableContentMaxWidthClassName } from "~/constants/ui"
 import { useRenderStyle } from "~/hooks/biz/useRenderStyle"
 import type { TextSelectionEvent } from "~/lib/simple-text-selection"
-import { useBlockActions } from "~/modules/ai-chat/__internal__/hooks"
-import { BlockSliceAction } from "~/modules/ai-chat/__internal__/slices/block.slice"
+import { useBlockActions } from "~/modules/ai-chat/store/hooks"
+import { BlockSliceAction } from "~/modules/ai-chat/store/slices/block.slice"
 import { EntryContentHTMLRenderer } from "~/modules/renderer/html"
 import { WrappedElementProvider } from "~/providers/wrapped-element-provider"
