+++ pnpm-workspace.yaml
@@ -58,4 +58,4 @@ overrides:
 
 catalog:
   typescript: "5.8.3"
-  "@follow-app/client-sdk": "0.3.25"
+  "@follow-app/client-sdk": "0.3.26"
